Added Curiosity "grouping" on ExtInventory
This is something that I wanted to add few days ago... ExtInventory add you the ability to group items regarding their Quality... but I found it a nice tool to do some other type grouping (kinda sorting for curios).

I always struggled to find the best curio (MentalWeight or Learning points per hour, LPH) on a cupboard/container fast. So I used the same logic for grouping to add two options>
 - Curiosities Mental Weight
 - Curiosities LPH

It will show you on the right "grouping" window a sorted version of each one of them. I know I'm stretching the idea of grouping, but this may be useful for other players :D or not...